Collocated (externally exposed) ReactTypes and ReactNativeTypes into single files to be synced to fbsource
