+++ scripts/rollup/shims/react-native/NativeMethodsMixin.js
@@ -16,5 +16,8 @@ const {
   __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,
 } = require('ReactNative');
 
-module.exports =
-  __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.NativeMethodsMixin;
+import type {NativeMethodsMixinType} from 'ReactNativeTypes';
+
+const {NativeMethodsMixin} = __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED;
+
+module.exports = ((NativeMethodsMixin: any): $Exact<NativeMethodsMixinType>);
