+++ src/renderers/shared/fiber/isomorphic/ReactPortal.js
@@ -12,22 +12,13 @@
 
 'use strict';
 
-import type {ReactNodeList} from 'ReactTypes';
+import type {ReactNodeList, ReactPortal} from 'ReactTypes';
 
 // The Symbol used to tag the special React types. If there is no native Symbol
 // nor polyfill, then a plain number is used for performance.
 var REACT_PORTAL_TYPE =
   (typeof Symbol === 'function' && Symbol.for && Symbol.for('react.portal')) ||
   0xeaca;
-
-export type ReactPortal = {
-  $$typeof: Symbol | number,
-  key: null | string,
-  containerInfo: any,
-  children: ReactNodeList,
-  // TODO: figure out the API for cross-renderer implementation.
-  implementation: any,
-};
 
 exports.createPortal = function(
   children: ReactNodeList,
